Question
1. Create a dataset containing information about the prices of houses in a particular city. The
dataset includes the following features for each house:
• Size (in square feet)
• Number of bedrooms
• Age of the house (in years)
• Distance to city center (in miles)
• Price (in $)
(a) Train a Linear Regression Model: Use the features (Size, Number of bedrooms, Age,
Distance to city center) to predict the price of houses.

(b) Evaluate the Model: After training the model on the training set, evaluate its perfor-
mance using Mean Squared Error (MSE) as performance metrics.

(c) Interpret the Results: Interpret the coefficients of the linear regression model. What
are the values of the coefficients for each feature in the model? What is the impact
of each feature on the predicted house price?
(d) Given the following house details, predict the price using your trained model:
• Size: 2500 sq. ft.
• Number of bedrooms: 4
• Age: 10 years
• Distance to city center: 5 miles